Inovar v1
This is the first version of another of my simple python projects that I did for evaluation purposes at school, this one is called "Inovar", which is one of two main systems for schools management, that we have here in Portugal.

I have experienced both GIAE and Inovar systems, in a students perspective and, I choose to do this project with this theme because my current school uses Inovar and one time when I was in one of my coding classes, our teacher told us that we all would do a tiny project for managing a school, with a generalist perspective. This one project only lasted for about 2-3 classes (with a total of about 3 to 6 hours), and I thought to myself well I already have a very generalist base which I can work around it, to create something close to what we have in our school but simpler, and then this project was born.

Please note that since Im from Portugal, and since Im still studying, I didnt quite had the time yet to adapt this project to a more universally understood language like English, and also note that this project has been done some time ago and that back then my method to declare variables was kinda weird, but unique.