Inovar v2 Final Version
This is the 2nd and the "last" version of my python project called: "Inovar"
If you read version 1 description you already have some background on the project if not, go check it out and come back after.

Changelog:
- Little optimization in the dictionary that lists every single student in the school;
- Update on os creating folders in a specified path (now should include a fully functional version for text files);
- Using try-except method to prevent errors in Login Screen;
- Little customization on welcome message;
- Update on system that assigns x student to its class (the system should now somehow be more efficient than the previous one);
- Big update on option 1.3 of the schools director menu, he should now be able to properly print to a .txt file info about a certain student;
- Big update on the second directors menu, now it fully suppports option 2.1 (Add grades) and option 2.2 (Remove grades).

As I said earlier this is the "last" version of my project because as mentioned before Im still studying and I dont have quite the time to update this old projects. After I can spend some time on adapting this project along with the others to English, I am considering maybe coding a next version that will be the real final version and maybe developing a version for linux/macOS users because so of the most vital features on the program are currently only available to Windows Users.